improved template-view

+++ src/app/templates/template-view/template-view.component.ts
@@ -22,7 +22,7 @@ export class TemplateViewComponent implements OnInit {
   Field = TemplateField;
   Table = TemplateTable;
 
-  template: Template = new Template();
+  template?: Template;
   templateTypes?: TemplateType[];
   private vacantId: number = 0;
   private id: number = -1;
@@ -37,31 +37,41 @@ export class TemplateViewComponent implements OnInit {
     this.templateSvc.getTemplate(this.id).subscribe(data => {
       this.template = data;
       this.vacantId = data.TemplateItems.length;
-      console.log(`received data: ${JSON.stringify(this.template, null, 2)}`);
     });
     this.typesSvc.getTypes().subscribe(data => {
       this.templateTypes = data;
-      console.log(`received types: ${JSON.stringify(this.templateTypes, null, 2)}`);
     });
   }
 
   changeUsage(){
+    if (!this.template) return;
+
     this.template.Depricated = !this.template.Depricated;
     this.save();
   }
 
   changeOrder(oldOrder: number, delta: number){
+    if (!this.template) return;
+
     let newOrder = oldOrder + delta;
     if (newOrder < this.template.TemplateItems.length && newOrder >= 0){
-      [this.template.TemplateItems[newOrder], this.template.TemplateItems[oldOrder]] = [this.template.TemplateItems[oldOrder], this.template.TemplateItems[newOrder]];
+      let id_1 = this.template.TemplateItems[oldOrder].Id;
+      let id_2 = this.template.TemplateItems[newOrder].Id;
+      this.templateSvc.moveItems(this.template.Id, id_1, id_2).subscribe(() => 
+        this.alertSvc.info("Поле перемещено", {single: true, closeTime: 5000})
+      );
+      [this.template.TemplateItems[newOrder], this.template.TemplateItems[oldOrder]] = 
+      [this.template.TemplateItems[oldOrder], this.template.TemplateItems[newOrder]];
     }
-    this.updateItem(this.template.TemplateItems[newOrder]);
-    this.updateItem(this.template.TemplateItems[oldOrder]);
   }
 
   addField(){
-    let field = new TemplateField(`Поле ${this.vacantId + 1}`, this.vacantId, this.template.Id);
+    if (!this.template) return;
+
+    let field = new TemplateField("Новое поле", this.vacantId, this.template.Id);
     this.templateSvc.updateField(this.template.Id, field).subscribe(f => {
+      if (!this.template) return;
+
       field.Id = f.Id;
       this.template.TemplateItems.push(field);    
       this.vacantId++;
@@ -70,30 +80,28 @@ export class TemplateViewComponent implements OnInit {
   }
 
   addColumn(column: TemplateField){
-    this.templateSvc.updateField(this.template.Id, column).subscribe({
-      next: () => this.alertSvc.info("Столбец сохранен"),
+    if (!this.template) return;
+
+    this.templateSvc.updateField(this.template.Id, column).subscribe(f => {
+      column.Id = f.Id;
+      this.alertSvc.info("Столбец сохранен");
     });
   }
 
   addTable(){
-    let table = new TemplateTable(this.template.Id,
-      `Таблица ${this.template.TemplateItems.length + 1}`,
-      1, this.vacantId);
-    let field = new TemplateField(`Столбец 1`, 0, this.template.Id, table.Id);
+    if (!this.template) return;
+
+    let table = new TemplateTable(this.template.Id, "Новая таблица", 1);
 
     this.templateSvc.updateTable(this.template.Id, table).subscribe(t => {
       table.Id = t.Id;
-      this.templateSvc.updateField(this.template.Id, field).subscribe(f => {
-        field.Id = f.Id;
-        this.template.TemplateItems.push(table);      
-          table.TemplateFields.push(field);
-          this.alertSvc.info("Таблица создана");
-          this.vacantId++;
-      });
+      this.alertSvc.info("Таблица создана");
     });
   }
 
   updateItem(item: TemplateField | TemplateTable){
+    if (!this.template) return;
+
     if (item.IsTable) {
       this.templateSvc.updateTable(this.template.Id, item).subscribe(() => 
         this.alertSvc.info("Таблица сохранена")
@@ -107,17 +115,23 @@ export class TemplateViewComponent implements OnInit {
   }
 
   deleteItem(item: ITemplateItem){
+    if (!this.template) return;
+
     this.templateSvc.deleteItem(this.template.Id, item.Id, item.IsTable).subscribe(() => {
+      if (!this.template) return;
+
       this.alertSvc.info(item.IsTable ? "Таблица удалена" : "Поле удалено");
       this.template.TemplateItems = this.template.TemplateItems.filter(i => i.Id != item.Id);
     });
   }
 
   save(){
+    if (!this.template) return;
     this.templateSvc.updateTemplate(this.template).subscribe(() => this.alertSvc.info("Шаблон сохранён"));
   }
 
   delete(){
+    if (!this.template) return;
     this.templateSvc.deleteTemplate(this.template.Id).subscribe(() => this.router.navigate(["/templates"]));
   }
 }
